+++ community/io/src/test/java/org/neo4j/io/pagecache/RecordingPageCacheTracer.java
@@ -101,7 +101,7 @@ public class RecordingPageCacheTracer implements PageCacheTracer
                 return new PageFaultEvent()
                 {
                     @Override
-                    public void addBytesRead( int bytes )
+                    public void addBytesRead( long bytes )
                     {
                     }
